+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.47.0] - 2023-08-04
+
 - Add `build_sdist` to get source distributions with `git-archive-all`
 - Split `cmeel.build.build` into multiple utils functions
 - Move `cmeel.build.build` to `cmeel.impl.build_impl`
@@ -355,6 +357,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial minial working example
 
 [Unreleased]: https://github.com/cmake-wheel/cmeel/compare/v0.4.2...main
+[v0.47.0]: https://github.com/cmake-wheel/cmeel/compare/v0.46.0...v0.47.0
 [v0.46.0]: https://github.com/cmake-wheel/cmeel/compare/v0.45.0...v0.46.0
 [v0.45.0]: https://github.com/cmake-wheel/cmeel/compare/v0.44.1...v0.45.0
 [v0.44.1]: https://github.com/cmake-wheel/cmeel/compare/v0.44.0...v0.44.1